后端

+++ src/main/java/com/springboot/service/impl/OrdersServiceImpl.java
@@ -3,46 +3,49 @@ package com.springboot.service.impl;
 
 import cn.hutool.core.util.IdUtil;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
-import com.baomidou.mybatisplus.core.metadata.OrderItem;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
-import com.springboot.controller.enums.OrderStatusEnum;
+
+import com.springboot.common.OrderStatusEnum;
 import com.springboot.entity.*;
 import com.springboot.exception.ServiceException;
 import com.springboot.mapper.OrdersMapper;
-import com.springboot.service.ICartService;
-import com.springboot.service.IGoodsService;
-import com.springboot.service.IItemService;
-import com.springboot.service.IOrdersService;
+import com.springboot.service.CartService;
+import com.springboot.service.GoodsService;
+import com.springboot.service.ItemService;
+import com.springboot.service.OrdersService;
 import com.springboot.utils.TokenUtils;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.annotation.Resource;
 import java.math.BigDecimal;
 import java.util.List;
+import java.util.Objects;
 
 
-/*
- * <p>
- *  服务实现类
- * </p>
+/**
+ * 服务实现类
  *
- * @author 
+ * @author
  * @since 2023-03-04
  */
 @Service
-public class OrdersServiceImpl extends ServiceImpl<OrdersMapper, Orders> implements IOrdersService {
+public class OrdersServiceImpl extends ServiceImpl<OrdersMapper, Orders> implements OrdersService {
 
-    @Autowired
-    private IItemService ItemService;
+    @Resource
+    private ItemService itemService;
 
-    @Autowired
-    private ICartService cartService;
+    @Resource
+    private CartService cartService;
 
-    @Autowired
-    IGoodsService goodsService;
+    @Resource
+    GoodsService goodsService;
 
-
+    /**
+     * 下订单
+     * @param carts   购物车勾选信息
+     * @param address 地址
+     */
     @Transactional
     @Override
     public void addOrder(List<Cart> carts,String address) {
@@ -75,9 +78,34 @@ public class OrdersServiceImpl extends ServiceImpl<OrdersMapper, Orders> impleme
             orderItem.setImg(cart.getImg());
             orderItem.setNum(cart.getNum());
             orderItem.setGoodsId(cart.getGoodsId());
-            ItemService.save(orderItem);
-            // 下单成功后删除购物车商品
-            cartService.removeById(cart.getId());
+            boolean result = itemService.save(orderItem);
+            if (result){
+                // 下单成功后删除购物车商品
+                cartService.removeById(cart.getId());
+            }
         }
     }
+
+
+    /**
+     * 修改订单状态
+     *
+     * @param orders
+     */
+    @Override
+    public Boolean updateOrder(Orders orders) {
+        if (Objects.equals(orders.getStatus(), OrderStatusEnum.CANCEL.getCode())) {
+            // 取消订单，加库存
+            Integer ordersId = orders.getId();
+            QueryWrapper<Item> order_id = new QueryWrapper<Item>().eq("order_id", ordersId);
+            List<Item> orderItems = itemService.list(order_id);
+            for (Item orderItem : orderItems) {
+                Goods goods = goodsService.getById(orderItem.getGoodsId());
+                goods.setNums(goods.getNums() + orderItem.getNum());
+                goodsService.updateById(goods);
+            }
+        }
+        return (updateById(orders));
+    }
+
 }
